Unify Asset and DDO classes.
