+++ ocean_lib/models/test/test_ddo.py
@@ -7,10 +7,9 @@ import uuid
 
 import pytest
 from eth_utils import add_0x_prefix, remove_0x_prefix
-from ocean_lib.assets.asset import Asset
+from ocean_lib.assets.asset import V3Asset
 from ocean_lib.common.agreements.consumable import ConsumableCodes, MalformedCredential
 from ocean_lib.common.ddo.credentials import AddressCredential
-from ocean_lib.common.ddo.ddo import DDO
 from ocean_lib.common.utils.utilities import checksum
 from ocean_lib.models.data_token import DataToken
 from ocean_lib.models.dtfactory import DTFactory
@@ -28,7 +27,7 @@ def get_ddo_sample(datatoken_address):
     sample_ddo_path = get_resource_path("ddo", "ddo_sa_sample.json")
     assert sample_ddo_path.exists(), "{} does not exist!".format(sample_ddo_path)
 
-    asset = DDO(json_filename=sample_ddo_path)
+    asset = V3Asset(json_filename=sample_ddo_path)
     asset.metadata["main"]["files"][0]["checksum"] = str(uuid.uuid4())
 
     checksum_dict = dict()
@@ -45,7 +44,7 @@ def test_ddo_credentials_addresses_both():
     sample_ddo_path = get_resource_path("ddo", "ddo_sa_sample_with_credentials.json")
     assert sample_ddo_path.exists(), "{} does not exist!".format(sample_ddo_path)
 
-    ddo = DDO(json_filename=sample_ddo_path)
+    ddo = V3Asset(json_filename=sample_ddo_path)
     address_credential = AddressCredential(ddo)
     assert address_credential.get_addresses_of_class("allow") == ["0x123", "0x456a"]
     assert address_credential.get_addresses_of_class("deny") == ["0x2222", "0x333"]
@@ -65,7 +64,7 @@ def test_ddo_credentials_addresses_only_deny():
     sample_ddo_path = get_resource_path("ddo", "ddo_sa_sample_with_credentials.json")
     assert sample_ddo_path.exists(), "{} does not exist!".format(sample_ddo_path)
     # remove allow to test the behaviour of deny
-    ddo = DDO(json_filename=sample_ddo_path)
+    ddo = V3Asset(json_filename=sample_ddo_path)
     ddo.credentials.pop("allow")
 
     address_credential = AddressCredential(ddo)
@@ -92,7 +91,7 @@ def test_ddo_credentials_addresses_no_access_list():
 
     # if "allow" OR "deny" exist, we need a credential,
     # so remove both to test the behaviour of no credential supplied
-    ddo = DDO(json_filename=sample_ddo_path)
+    ddo = V3Asset(json_filename=sample_ddo_path)
     address_credential = AddressCredential(ddo)
     ddo.credentials.pop("allow")
     ddo.credentials.pop("deny")
@@ -109,7 +108,7 @@ def test_ddo_credentials_addresses_no_access_list():
 
 
 def test_ddo_connection(config):
-    ddo = DDO("did:op:testdid")
+    ddo = V3Asset("did:op:testdid")
     provider_uri = config.provider_url
     assert (
         ddo.is_consumable(with_connectivity_check=True, provider_uri=provider_uri)
@@ -121,7 +120,7 @@ def test_ddo_credentials_disabled():
     sample_ddo_path = get_resource_path("ddo", "ddo_sa_sample_disabled.json")
     assert sample_ddo_path.exists(), "{} does not exist!".format(sample_ddo_path)
 
-    ddo = DDO(json_filename=sample_ddo_path)
+    ddo = V3Asset(json_filename=sample_ddo_path)
     assert ddo.is_disabled
     assert not ddo.is_enabled
 
@@ -165,7 +164,7 @@ def test_ddo_on_chain(config, web3):
     ddo_text = Web3.toText(lzma.decompress(log.args.data))
     assert ddo_text == asset.as_text(), "ddo text does not match original."
 
-    _asset = Asset(json_text=ddo_text)
+    _asset = V3Asset(json_text=ddo_text)
     assert _asset.did == asset.did, "did does not match."
     name = _asset.metadata["main"]["name"]
     assert name == old_name, f"name does not match: {name} != {old_name}"
@@ -185,7 +184,7 @@ def test_ddo_on_chain(config, web3):
     # read back the asset ddo from the event log
     ddo_text = Web3.toText(lzma.decompress(log.args.data))
     assert ddo_text == asset.as_text(), "ddo text does not match original."
-    _asset = Asset(json_text=ddo_text)
+    _asset = V3Asset(json_text=ddo_text)
     assert (
         _asset.metadata["main"]["name"] == "updated name for test"
     ), "name does not seem to be updated."
@@ -218,7 +217,7 @@ def test_ddo_address_utilities():
     sample_ddo_path = get_resource_path("ddo", "ddo_sa_sample_with_credentials.json")
     assert sample_ddo_path.exists(), "{} does not exist!".format(sample_ddo_path)
 
-    ddo = DDO(json_filename=sample_ddo_path)
+    ddo = V3Asset(json_filename=sample_ddo_path)
 
     assert ddo.allowed_addresses == ["0x123", "0x456a"]
 
@@ -238,7 +237,7 @@ def test_ddo_address_utilities():
     ddo.add_address_to_deny_list("0xasfaweg")
     assert ddo.denied_addresses == ["0x2222", "0x333", "0xasfaweg"]
 
-    ddo = DDO()
+    ddo = V3Asset()
     assert ddo.allowed_addresses == []
     ddo.add_address_to_allow_list("0xAbc12")
     assert ddo.allowed_addresses == ["0xabc12"]
@@ -251,7 +250,7 @@ def test_ddo_retiring():
     sample_ddo_path = get_resource_path("ddo", "ddo_sa_sample.json")
     assert sample_ddo_path.exists(), "{} does not exist!".format(sample_ddo_path)
 
-    ddo = DDO(json_filename=sample_ddo_path)
+    ddo = V3Asset(json_filename=sample_ddo_path)
     assert not ddo.is_retired
 
     ddo.retire()
@@ -266,7 +265,7 @@ def test_ddo_unlisting():
     sample_ddo_path = get_resource_path("ddo", "ddo_sa_sample.json")
     assert sample_ddo_path.exists(), "{} does not exist!".format(sample_ddo_path)
 
-    ddo = DDO(json_filename=sample_ddo_path)
+    ddo = V3Asset(json_filename=sample_ddo_path)
     assert ddo.is_listed
 
     ddo.unlist()
