+++ ocean_lib/test/test_exceptions.py
@@ -6,7 +6,7 @@ import uuid
 from unittest.mock import patch
 
 import pytest
-from ocean_lib.common.ddo.ddo import DDO
+from ocean_lib.assets.asset import V3Asset
 from ocean_lib.exceptions import AquariusError, ContractNotFound, InsufficientBalance
 from ocean_lib.web3_internal.constants import ZERO_ADDRESS
 from ocean_lib.web3_internal.currency import to_wei
@@ -17,7 +17,7 @@ from tests.resources.helper_functions import get_publisher_wallet
 def test_InsufficientBalance(publisher_ocean_instance):
     alice = get_publisher_wallet()
     sample_ddo_path = get_resource_path("ddo", "ddo_sa_sample.json")
-    asset = DDO(json_filename=sample_ddo_path)
+    asset = V3Asset(json_filename=sample_ddo_path)
     asset.metadata["main"]["files"][0]["checksum"] = str(uuid.uuid4())
 
     token = publisher_ocean_instance.create_data_token(
